creating a vmath structure for MATLAB-style vectorized math functions
